requirements.txt
Cython==0.29.1
imageio==2.1.2
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
notebook==5.7.2
numpy==1.15.4
ptyprocess==0.6.0
py==1.8.0
pybind11==2.2.4
pycrypto==2.6.1
six==1.11.0
tornado==5.1.1
tqdm==4.11.2
traitlets==4.3.2
typing==3.6.6
Unidecode==1.0.23
urllib3==1.22
webencodings==0.5.1
